[composite compliance] fix max_pool1d
max_pool1d has a fast path for CPU tensors that do not require grad that
directly accesses the data_ptr. This PR makes the change that if the
input Tensor is a Tensor Subclass, then we want to walk through the
"slow path" of calling max_pool1d_with_indices.

Test Plan:
- wait for tests

ghstack-source-id: 2a26b1ab0ceac6ec32ae17d34126871130bd8b99
Pull Request resolved: https://github.com/pytorch/pytorch/pull/84127